Mental_health_chatbot/requirements.txt
# core libraries
langchain==0.1.3
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=0.27.0
llama-index
haystack

#for a backend server 
fastapi
uvicorn

#additional tools
requests
python-dotenv

#optional ai agent frameworks
autogpt
